Fixup configuration and move out of handlers.
(should make Grim happy)
