+++ settings/configuration.go
@@ -1,4 +1,4 @@
-package handlers
+package settings
 
 import (
 	"io/ioutil"
@@ -9,7 +9,31 @@ import (
 	"gopkg.in/yaml.v2"
 )
 
-func setDefaults(config *types.Config) {
+var config types.Config
+
+func init() {
+	// If no config file is found set defaults.
+	configfile, err := filepath.Abs("./config.yml")
+	if err != nil {
+		log.Println(err)
+		setDefaults()
+		return
+	}
+	// If unable to read the file set defaults.
+	configyaml, err := ioutil.ReadFile(configfile)
+	if err != nil {
+		log.Println(err)
+		setDefaults()
+		return
+	}
+	// If unable to parse the yaml set defaults.
+	if err := yaml.Unmarshal([]byte(configyaml), &config); err != nil {
+		log.Println(err)
+		setDefaults()
+	}
+}
+
+func setDefaults() {
 	config.Port = ":4002"
 	// TODO benchmark and tune default buffer-sizes
 	config.ReadBufferSize = 1024
@@ -19,25 +43,7 @@ func setDefaults(config *types.Config) {
 	config.MessageOverflow = 0
 }
 
-// InitConfig should be ran during start to load all server configuration data.
-func InitConfig(config *types.Config) {
-	// If no config file is found set defaults.
-	configfile, err := filepath.Abs("./config.yml")
-	if err != nil {
-		log.Println(err)
-		setDefaults(config)
-		return
-	}
-	// If unable to read the file set defaults.
-	configyaml, err := ioutil.ReadFile(configfile)
-	if err != nil {
-		log.Println(err)
-		setDefaults(config)
-		return
-	}
-	// If unable to parse the yaml set defaults.
-	if err := yaml.Unmarshal([]byte(configyaml), &config); err != nil {
-		log.Println(err)
-		setDefaults(config)
-	}
+// GetConfig returns the current configuration file.
+func GetConfig() types.Config {
+	return config
 }
